ECF-RP-315: refactor admin supplier routes
