+++ app/views/admin/suppliers/delivery_partners/delivery_partner_success.html.erb
