+++ app/views/admin/suppliers/delivery_partners/review_delivery_partner.html.erb
@@ -1,6 +1,6 @@
 <% content_for :before_content, govuk_back_link(
   text: "Back",
-  href: admin_new_lead_provider_cohorts_path)
+  href: choose_lps_admin_delivery_partners_path)
 %>
 
 <h1 class="govuk-heading-l">Confirm these details</h1>
@@ -11,7 +11,7 @@
       Name
     </dt>
     <dd class="govuk-summary-list__value">
-      <%= @lead_provider_form.name %>
+      <%= @delivery_partner_form.name %>
     </dd>
     <dd class="govuk-summary-list__actions">
       <%= govuk_link_to new_admin_supplier_path(continue: true), class: "govuk-link--no-visited-state" do %>
@@ -24,44 +24,36 @@
       Type
     </dt>
     <dd class="govuk-summary-list__value">
-      Lead Provider
+      Delivery Partner
     </dd>
     <dd class="govuk-summary-list__actions">
-      <%= govuk_link_to admin_new_supplier_type_path, class: "govuk-link--no-visited-state" do %>
+      <%= govuk_link_to supplier_type_admin_suppliers_path, class: "govuk-link--no-visited-state" do %>
         Change<span class="govuk-visually-hidden"> supplier type</span>
       <% end %>
     </dd>
   </div>
   <div class="govuk-summary-list__row">
     <dt class="govuk-summary-list__key">
-      Course
+      Working with
     </dt>
     <dd class="govuk-summary-list__value">
-      <%= @lead_provider_form.chosen_cip.name %>
-    </dd>
-    <dd class="govuk-summary-list__actions">
-      <%= govuk_link_to admin_new_lead_provider_cip_path, class: "govuk-link--no-visited-state" do %>
-        Change<span class="govuk-visually-hidden"> course</span>
+      <% @delivery_partner_form.display_lead_provider_details.each do |lead_provider| %>
+      <strong><%= lead_provider[:name] %></strong>
+        <br/>
+        In cohort(s) <%= lead_provider[:chosen_cohorts] %>
+        <br/>
       <% end %>
-    </dd>
-  </div>
-  <div class="govuk-summary-list__row">
-    <dt class="govuk-summary-list__key">
-      Cohort(s)
-    </dt>
-    <dd class="govuk-summary-list__value">
-      <%= @lead_provider_form.chosen_cohort_names %>
 
     </dd>
     <dd class="govuk-summary-list__actions">
-      <%= govuk_link_to admin_new_lead_provider_cohorts_path, class: "govuk-link--no-visited-state" do %>
-        Change<span class="govuk-visually-hidden"> cohorts</span>
+      <%= govuk_link_to choose_lps_admin_delivery_partners_path, class: "govuk-link--no-visited-state" do %>
+        Change<span class="govuk-visually-hidden"> lead providers and cohorts</span>
       <% end %>
     </dd>
   </div>
 </dl>
 
-<%= form_for :confirmation, url: admin_create_lead_provider_path, method: :post do |f| %>
+<%= form_for :confirmation, url: admin_create_delivery_partner_path, method: :post do |f| %>
   <%= f.hidden_field :answers_checked, value: true %>
   <%= f.govuk_submit "Create supplier" %>
 <% end %>
