+++ app/views/admin/suppliers/lead_providers/choose_cohorts.html.erb
@@ -1,9 +1,9 @@
 <% content_for :before_content, govuk_back_link(
   text: "Back",
-  href: admin_new_lead_provider_cip_path)
+  href: choose_cip_admin_lead_providers_path)
 %>
 
-<%= form_with model: @lead_provider_form, url: admin_new_lead_provider_cohorts_path, method: :post do |f| %>
+<%= form_with model: @lead_provider_form, url: choose_cohorts_admin_lead_providers_path, method: :post do |f| %>
   <%= f.govuk_collection_check_boxes(
         :cohorts,
         @lead_provider_form.available_cohorts,
