+++ app/views/admin/suppliers/supplier_users/index.html.erb
@@ -1,4 +1,4 @@
-<%= render "admin/suppliers/layout", locals = { page: "users" } %>
+<%= render "admin/suppliers/suppliers/layout", locals = { page: "users" } %>
 <%= govuk_link_to "Add a new user", new_admin_supplier_user_path, button: true %>
 <table class="govuk-table">
   <thead>
