+++ app/views/admin/suppliers/suppliers/new.html.erb
@@ -1,7 +1,7 @@
 <% content_for :before_content, govuk_back_link(text: "Back", href: admin_suppliers_path) %>
 
 <%= @value %>
-<%= form_with model: @new_supplier_form, url: admin_new_supplier_path, method: :post do |f| %>
+<%= form_with model: @new_supplier_form, url: admin_suppliers_path, method: :post do |f| %>
   <%= f.govuk_text_field(
         :name,
         width: "three-quarters",
