+++ config/routes.rb
@@ -57,60 +57,47 @@ Rails.application.routes.draw do
   end
 
   namespace :admin do
-    resources :suppliers, only: %i[index new]
-    scope "suppliers/new" do
-      post "/", controller: :suppliers, action: :receive_new, as: :new_supplier
-      get "supplier-type", controller: :suppliers, action: :new_supplier_type, as: :new_supplier_type
-      post "supplier-type", controller: :suppliers, action: :receive_new_supplier_type
+    scope :suppliers, module: "suppliers" do
+      resources :suppliers, only: %i[index new], path: "/"
+      scope "new" do
+        resources :suppliers, only: :create, path: "/" do
+          collection do
+            get "supplier-type", action: :new_supplier_type
+            post "supplier-type", action: :receive_new_supplier_type
+          end
+        end
 
-      scope "lead-provider" do
-        get "choose-cip", controller: :lead_providers, action: :choose_cip, as: :new_lead_provider_cip
-        post "choose-cip", controller: :lead_providers, action: :receive_cip
-        get "choose-cohorts", controller: :lead_providers, action: :choose_cohorts, as: :new_lead_provider_cohorts
-        post "choose-cohorts", controller: :lead_providers, action: :receive_cohorts
-        get "review", controller: :lead_providers, action: :review, as: :new_lead_provider_review
-        post "/", controller: :lead_providers, action: :create, as: :create_lead_provider
-        get "success", controller: :lead_providers, action: :success, as: :new_lead_provider_success
+        resources :lead_providers, only: [], path: "lead-provider" do
+          collection do
+            get "choose-cip", action: :choose_cip
+            post "choose-cip", action: :receive_cip
+            get "choose-cohorts", action: :choose_cohorts
+            post "choose-cohorts", action: :receive_cohorts
+            get "review", action: :review
+            post "/", action: :create
+            get "success", action: :success
+          end
+        end
+
+        resources :delivery_partners, only: [], path: "delivery-partner" do
+          collection do
+            get "choose-lps", action: :choose_lead_providers
+            post "choose-lps", action: :receive_lead_providers
+            get "review", action: :review_delivery_partner
+            post "/", action: :create_delivery_partner
+            get "success", action: :delivery_partner_success
+          end
+        end
       end
 
-      scope "delivery-partner" do
-        get "choose-lps", controller: :delivery_partners, action: :choose_lead_providers, as: :new_delivery_partner_lps
-        post "choose-lps", controller: :delivery_partners, action: :receive_lead_providers
-        get "review", controller: :delivery_partners, action: :review_delivery_partner, as: :new_delivery_partner_review
-        post "/", controller: :delivery_partners, action: :create_delivery_partner, as: :create_delivery_partner
-        get "success", controller: :delivery_partners, action: :delivery_partner_success, as: :new_delivery_partner_success
+      resources :supplier_users, only: %i[index new create], path: "users"
+      scope "users/new" do
+        post "/", controller: :supplier_users, action: :receive_supplier
+        get "user-details", controller: :supplier_users, action: :user_details, as: :new_supplier_user_details
+        post "user-details", controller: :supplier_users, action: :receive_user_details
+        get "review", controller: :supplier_users, action: :review, as: :new_supplier_user_review
+        get "success", controller: :supplier_users, action: :success, as: :new_supplier_user_success
       end
-    end
-
-    scope "lead-providers/:lead_provider" do
-      get "/", controller: :lead_providers, action: :show_details, as: :show_lead_provider
-      get "/users", controller: :lead_providers, action: :show_users, as: :show_lead_provider_users
-      get "/delivery-partners", controller: :lead_providers, action: :show_dps, as: :show_lead_provider_dps
-      get "/schools", controller: :lead_providers, action: :show_schools, as: :show_lead_provider_schools
-
-      get "/edit", controller: :lead_providers, action: :edit, as: :edit_lead_provider
-      post "/", controller: :suppliers, action: :update_lead_provider, as: :update_lead_provider
-      resources :lead_provider_users, path: "/users"
-    end
-
-    scope "delivery-partners/:delivery_partner" do
-      get "/", controller: :delivery_partners, action: :show_users, as: :show_delivery_partner
-      get "/lead-providers", controller: :delivery_partners, action: :show_lps, as: :show_delivery_partner_lps
-      get "/schools", controller: :delivery_partners, action: :show_schools, as: :show_delivery_partner_schools
-
-      get "/edit", controller: :delivery_partners, action: :edit_delivery_partner, as: :edit_delivery_partner
-      post "/", controller: :delivery_partners, action: :update_delivery_partner, as: :update_delivery_partner
-
-      resources :lead_provider_users, path: "/users"
-    end
-
-    resources :supplier_users, only: %i[index new create], path: "suppliers/users"
-    scope "suppliers/users/new" do
-      post "/", controller: :supplier_users, action: :receive_supplier
-      get "user-details", controller: :supplier_users, action: :user_details, as: :new_supplier_user_details
-      post "user-details", controller: :supplier_users, action: :receive_user_details
-      get "review", controller: :supplier_users, action: :review, as: :new_supplier_user_review
-      get "success", controller: :supplier_users, action: :success, as: :new_supplier_user_success
     end
   end
